Railway architecture created
