--- lalala.txt_copy
@@ -1,2 +1,0 @@
-
-
