--- src/main/java/Alex/FileHelper/RunFileFilter.java
@@ -1,15 +1,0 @@
-package Alex.FileHelper;
-
-import java.io.File;
-import java.io.FileNotFoundException;
-
-/**
- * Created by alex on 16.11.15.
- */
-public class RunFileFilter {
-
-    public static void main (String []args){
-
-        new View().start();
-    }
-}
