--- src/main/java/Alex/hwWeek1/RunCalculator.java
@@ -1,14 +1,0 @@
-package Alex.hwWeek1;
-
-/**
- * Created by alex on 28.11.15.
- */
-public class RunCalculator {
-
-    public static void main (String []args){
-
-       new CalculatorImpl();
-
-
-    }
-}
