--- src/main/java/Alex/hwWeek2/exceptions/NullSubjectException.java
@@ -1,10 +1,0 @@
-package Alex.hwWeek2.exceptions;
-
-/**
- * Created by alex on 30.11.15.
- */
-public class NullSubjectException extends Throwable {
-
-
-    //public NullSubjectException(String message){super (message);}
-}
