--- src/main/java/Nata/Nata_Start
@@ -1,11 +1,0 @@
-package Natalie;
-
-/**
- * Created by Family on 11/15/2015.
- */
-public class Start {
-   public static void main (String []args){
-
-       new View().start();
-   }
-}
