+++ src/main/java/alex/fileHelper/RunFileFilter.java
@@ -1,0 +1,12 @@
+package alex.fileHelper;
+
+/**
+ * Created by alex on 16.11.15.
+ */
+public class RunFileFilter {
+
+    public static void main (String []args){
+
+        new View().start();
+    }
+}
