+++ src/main/java/alex/fileHelper/Utils.java
@@ -1,0 +1,75 @@
+package alex.fileHelper;
+
+import java.io.*;
+
+/**
+ * Created by alex on 20.11.15.
+ */
+public class Utils <T> implements Serializable{
+
+    public void save(String path, T object) {
+        try (ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream(path))) {
+            oos.writeObject(object);
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    public String fileReader (String path){
+        String fileContent ="";
+        BufferedReader br = null;
+        try {
+            br = new BufferedReader(new FileReader(path));
+            StringBuilder sb = new StringBuilder();
+            String line = br.readLine();
+            while (line != null) {
+                sb.append(line);
+                sb.append(System.lineSeparator());
+                line = br.readLine();
+            }
+            fileContent = sb.toString();
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+        } catch (IOException e) {
+            e.printStackTrace();
+        } finally {
+            try {
+                br.close();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+        return fileContent ;
+    }
+
+
+    public void fileWriter(File file, String content){
+        PrintWriter writer = null;
+        try {
+            writer = new PrintWriter(file);
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+        }
+        writer.println(content);
+        writer.close();
+    }
+
+    public T load(String path) {
+
+        try (ObjectInputStream ois = new ObjectInputStream(new FileInputStream(path))) {
+            return (T) ois.readObject();
+        } catch (IOException e) {
+            e.printStackTrace();
+        } catch (ClassNotFoundException e) {
+            e.printStackTrace();
+        }
+
+        return null;
+    }
+
+
+
+
+}
