+++ src/main/java/alex/hwWeek1/RunCalculator.java
@@ -1,0 +1,14 @@
+package alex.hwWeek1;
+
+/**
+ * Created by alex on 28.11.15.
+ */
+public class RunCalculator {
+
+    public static void main (String []args){
+
+       new CalculatorImpl();
+
+
+    }
+}
