+++ src/main/java/alex/hwWeek2/exceptions/NullStudentException.java
@@ -1,0 +1,12 @@
+package alex.hwWeek2.exceptions;
+
+/**
+ * Created by alex on 30.11.15.
+ */
+public class NullStudentException extends Exception {
+
+
+
+   // public NullStudentException(String message){super(message);}
+
+}
