+++ src/main/java/alex/hwWeek2/exceptions/NullSubjectException.java
@@ -1,0 +1,10 @@
+package alex.hwWeek2.exceptions;
+
+/**
+ * Created by alex on 30.11.15.
+ */
+public class NullSubjectException extends Throwable {
+
+
+    //public NullSubjectException(String message){super (message);}
+}
