+++ src/main/java/alex/ticketStore/dao/TicketDao.java
@@ -1,0 +1,7 @@
+package alex.ticketStore.dao;
+
+/**
+ * Created by alex on 03.12.15.
+ */
+public interface TicketDao {
+}
