+++ src/main/java/alex/ticketStore/enteties/Ticket.java
@@ -1,0 +1,7 @@
+package alex.ticketStore.enteties;
+
+/**
+ * Created by alex on 03.12.15.
+ */
+public class Ticket {
+}
