+++ src/main/java/alex/ticketStore/service/RailwayStation.java
@@ -1,0 +1,7 @@
+package alex.ticketStore.service;
+
+/**
+ * Created by alex on 03.12.15.
+ */
+public interface RailwayStation {
+}
